DX-1071: Move RSpec matchers to matchers directory
